requirements.txt
pytest
pytest-cov
fastapi
uvicorn
requests
python-dotenv
jinja2
